+++ conjure-java-core/src/test/java/com/palantir/conjure/java/parameterized/objects/FilePath.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.palantir.conjure.java.parameterized;
+package com.palantir.conjure.java.parameterized.objects;
 
 import com.fasterxml.jackson.annotation.JsonCreator;
 import com.fasterxml.jackson.annotation.JsonValue;
